% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{add_argument}
\alias{add_argument}
\title{Add argument to a query}
\usage{
add_argument(argument_values, argument_name, type, query, check = TRUE)
}
\arguments{
\item{argument_values}{value of the argument}

\item{argument_name}{name of the argument}

\item{type}{argument type from metadata}

\item{query}{existing query to append to}

\item{check}{if to check again metadata values}
}
\description{
Add argument to a query
}

