% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{check_ingredients}
\alias{check_ingredients}
\title{Check ingredients}
\usage{
check_ingredients(ingredients)
}
\arguments{
\item{ingredients}{ingridients to check}
}
\description{
Check ingredients list against predifined ingredients by Yummly
}
\note{
Predifined list is downloaded from Metadata Dictionaries
}

