% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{perform_query}
\alias{perform_query}
\title{Process query}
\usage{
perform_query(query)
}
\arguments{
\item{query}{string query to execute}
}
\description{
Query Yummly API and check return codes
}

