% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{prepare_array_parameter}
\alias{prepare_array_parameter}
\title{Prepare search parameter}
\usage{
prepare_array_parameter(param, name)
}
\arguments{
\item{param}{vector parameter to use}

\item{name}{name for parameter to use}
}
\description{
Prepare search parameter from direction
}

