% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{reassign_env}
\alias{reassign_env}
\title{Reassing object in the environment}
\usage{
reassign_env(name, obj, env)
}
\arguments{
\item{name}{name of an object to be replaced}

\item{obj}{object that will be put in the environment}

\item{env}{environment to be replaced in}
}
\description{
Wrapper to reassign function in namespace that deals with lockedBinding if needed
}

