% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{save_yummly_credentials}
\alias{save_yummly_credentials}
\title{Save API credentials for later use}
\usage{
save_yummly_credentials(app_id, app_key)
}
\arguments{
\item{app_id}{application ID}

\item{app_key}{application key}
}
\description{
This functions caches the credentials to avoid need for entering it when calling other functions
}
\examples{
# since not checking is preformed not to waste API calls
# it falls on the user to save correct information
save_yummly_credentials("APP_ID", "APP_KEY")
}

