% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{search_recipes}
\alias{search_recipes}
\title{Search recipes on Yummly.com}
\usage{
search_recipes(search_words, require_pictures, allowed_ingredient,
  excluded_ingredient, allowed_diet, allowed_allergy, allowed_cuisine,
  excluded_cuisine, allowed_course, excluded_course, allowed_holiday,
  excluded_holiday, max_total_time, max_results, start, nutrition, flavor,
  facet_field, app_id = auth_cache$APP_ID, app_key = auth_cache$APP_KEY)
}
\arguments{
\item{search_words}{search phrase, can be supplied in from of vector of words}

\item{require_pictures}{set to \code{TRUE} if only to return recipes with photos}

\item{allowed_ingredient}{ingredient that all search results must include}

\item{excluded_ingredient}{ingredient that all search results should not contain}

\item{allowed_diet}{search results will only include recipes whose ingredients are allowed for that diet}

\item{allowed_allergy}{only include recipes whose ingredients are allowed for that allergy}

\item{allowed_cuisine}{search results will only include recipes with that cuisine}

\item{excluded_cuisine}{search results will only exclude recipes with that cuisine}

\item{allowed_course}{search results will only include recipes with that cuisine}

\item{excluded_course}{search results will only exclude recipes with that cuisine}

\item{excluded_holiday}{search results will only exclude recipes with that holiday}

\item{max_total_time}{search for recipes that do not exceed a specified max total cook + prep time in seconds}

\item{max_results}{number of results to return}

\item{start}{start with specific result in search}

\item{nutrition}{set the range of allowed values for a given nutrition attribute (see below for the list of supported nutrition attributes) by setting a min and/or a max}

\item{flavor}{set the ranges for taste attributes (this corresponds to the taste sliders on the Yummly.com search page). The values of min and max are between 0 and 1.}

\item{facet_field}{facet counts for ingredient and diet. When this parameter is called, the response will include a facetCounts object that lists the matching diets or ingredients and how many results match each diet or ingredient.}

\item{app_id}{application ID}

\item{app_key}{application key}

\item{allowed}{holiday search results will only include recipes with that holiday}
}
\description{
Query Yummly.com API to search for recipes with certain parameter. All parameters are optional and can be used in any combination.
The criteria you pass via the various parameters are combined with the AND operator (set conjunction). In other words, every recipe has to match the search phrase and satisfy the ingredient, cuisine, course, holiday, time, nutrition, and taste restrictions as described below.
If you specify a multi-word phrase to the q parameter, every word has to match something in each matching recipe:
}
\note{
This function resembles search query to Yummly API
}
\references{
\itemize{
  \item Yummly Developer Guide \url{https://developer.yummly.com/documentation}
}
}

