% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{check_arguments}
\alias{check_arguments}
\title{Check ingredients}
\usage{
check_arguments(arguments, type)
}
\arguments{
\item{arguments}{arguments from metadata to check to check}

\item{type}{type of arguments from metadata}
}
\description{
Check ingredients list against predifined ingredients by Yummly
}
\note{
Predifined list is downloaded from Metadata Dictionaries
}

