% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.R
\name{get_recipe}
\alias{get_recipe}
\title{Get recipe from Yummly.com}
\usage{
get_recipe(recipe_id, app_id = auth_cache$APP_ID,
  app_key = auth_cache$APP_KEY)
}
\arguments{
\item{recipe_id}{recipe ID}

\item{app_id}{application ID}

\item{app_key}{application KEY}

\item{app_key}{application key}
}
\description{
This call is equivalent in functionality to a Yummly recipe page.
}
\note{
This function resembles viewing a recipe on Yummly.com
}
\examples{
\dontrun{
# to request the response for French Onion Soup by Ree Drummond The Pioneer Woman with id French-Onion-Soup-The-Pioneer-Woman-Cooks-_-Ree-Drummond-41364
get_recipe("French-Onion-Soup-The-Pioneer-Woman-Cooks-_-Ree-Drummond-41364")
}
}
\references{
\itemize{
  \item Yummly Developer Guide \url{https://developer.yummly.com/documentation}
}
}

