% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_jsonp}
\alias{parse_jsonp}
\title{Parse JSONP returned by Yummly for metadata}
\usage{
parse_jsonp(jsonp)
}
\arguments{
\item{jsonp}{jsonp string}
}
\description{
This function parses JSONP that yummly uses as a response.
It is based on assumption that list of elements is returned.
}

