% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{search_recipes}
\alias{search_recipes}
\title{Search recipes on Yummly.com}
\usage{
search_recipes(search_words, require_pictures, allowed_ingredient,
  excluded_ingredient, allowed_diet, allowed_allergy, allowed_cuisine,
  excluded_cuisine, allowed_course, excluded_course, allowed_holiday,
  excluded_holiday, max_total_time, max_results, start, nutrition, flavor,
  facet_field, app_id = auth_cache$APP_ID, app_key = auth_cache$APP_KEY)
}
\arguments{
\item{search_words}{search phrase, can be supplied in from of vector of words}

\item{allowed_ingredient}{ingredient that all search results must include}

\item{excluded_ingredient}{ingredient that all search results should not contain}

\item{app_id}{application ID}

\item{app_key}{application key}
}
\description{
Query Yummly.com API to search for recipes with certain parameters
}
\note{
This function resembles search query to Yummly API
}
\references{
\itemize{
  \item Yummly Developer Guide \url{https://developer.yummly.com/documentation}
}
}

